perf(ui): one request per page of snapshots, not three
The conversations were read inside the snapshots query, so every page turn, every
sort and every keystroke re-read them alongside the page: turning a page cost
three requests and typing a five-letter search cost fifteen. They do not depend on
the query — they only decide which names link anywhere — so they now have their own
key and are read once.

The filter is debounced for the same reason, now that it travels to the controller.
`useDebounced` moves out of SubstratePage, which had it for this exact problem.

Also restores the page's empty state: the shared rule reads a `{snapshots, total}`
object as "not empty", so "No snapshots yet" could never show.
